Move the brand design update toggle into the design system
