+++ app/src/main/java/com/duckduckgo/app/branddesignupdate/AppBrandDesignUpdatePreWarmObserver.kt
@@ -19,6 +19,7 @@ package com.duckduckgo.app.branddesignupdate
 import androidx.lifecycle.LifecycleOwner
 import com.duckduckgo.app.di.AppCoroutineScope
 import com.duckduckgo.app.lifecycle.MainProcessLifecycleObserver
+import com.duckduckgo.common.ui.store.AppBrandDesignUpdateToggles
 import com.duckduckgo.common.utils.DispatcherProvider
 import com.duckduckgo.di.scopes.AppScope
 import com.squareup.anvil.annotations.ContributesMultibinding
