--- app/src/main/java/com/duckduckgo/app/branddesignupdate/AppBrandDesignUpdateThemeProvider.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright (c) 2026 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.app.branddesignupdate
-
-import com.duckduckgo.common.ui.store.AppBrandDesignThemeProvider
-import com.duckduckgo.di.scopes.AppScope
-import com.squareup.anvil.annotations.ContributesBinding
-import javax.inject.Inject
-
-@ContributesBinding(AppScope::class)
-class AppBrandDesignUpdateThemeProvider @Inject constructor(
-    private val toggles: AppBrandDesignUpdateToggles,
-) : AppBrandDesignThemeProvider {
-
-    override fun isAppBrandDesignUpdateEnabled(): Boolean = toggles.feature().isEnabled()
-}
