+++ app/src/test/java/com/duckduckgo/app/branddesignupdate/AppBrandDesignUpdatePreWarmObserverTest.kt
@@ -17,6 +17,7 @@
 package com.duckduckgo.app.branddesignupdate
 
 import com.duckduckgo.common.test.CoroutineTestRule
+import com.duckduckgo.common.ui.store.AppBrandDesignUpdateToggles
 import com.duckduckgo.common.utils.DispatcherProvider
 import com.duckduckgo.feature.toggles.api.Toggle
 import kotlinx.coroutines.ExperimentalCoroutinesApi
